visibilityをVisibilityContainerに分離
